--- alembic/versions/4a7172864e2f_added_price_columns_and_minor_table_.py
@@ -1,50 +1,0 @@
-"""added price columns and minor table relationship changes
-
-Revision ID: 4a7172864e2f
-Revises: 8b76cfc4f4b7
-Create Date: 2024-09-04 21:47:06.230891
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '4a7172864e2f'
-down_revision: Union[str, None] = '8b76cfc4f4b7'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('attractions', sa.Column('price', sa.Integer(), server_default=sa.text('0'), nullable=True))
-    op.add_column('hotels_and_restaurants', sa.Column('price', sa.Integer(), server_default=sa.text('0'), nullable=True))
-    op.alter_column('itineraries', 'user_itinerary_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    op.alter_column('transportation', 'origin',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    op.alter_column('transportation', 'destination',
-               existing_type=sa.VARCHAR(),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('transportation', 'destination',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.alter_column('transportation', 'origin',
-               existing_type=sa.VARCHAR(),
-               nullable=False)
-    op.alter_column('itineraries', 'user_itinerary_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    op.drop_column('hotels_and_restaurants', 'price')
-    op.drop_column('attractions', 'price')
-    # ### end Alembic commands ###
