--- alembic/versions/8b76cfc4f4b7_added_relationship_tables_for_itinerary_.py
@@ -1,91 +1,0 @@
-"""added relationship tables for itinerary and categories
-
-Revision ID: 8b76cfc4f4b7
-Revises: c00423de709a
-Create Date: 2024-09-02 00:31:32.575577
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '8b76cfc4f4b7'
-down_revision: Union[str, None] = 'c00423de709a'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('user_itineraries',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('itineraries',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('user_itinerary_id', sa.Integer(), nullable=True),
-    sa.Column('district_id', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['district_id'], ['districts.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_itinerary_id'], ['user_itineraries.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('itinerary_activities',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('itinerary_id', sa.Integer(), nullable=False),
-    sa.Column('activity_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['activity_id'], ['activities.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['itinerary_id'], ['itineraries.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('itinerary_attractions',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('itinerary_id', sa.Integer(), nullable=False),
-    sa.Column('attraction_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['attraction_id'], ['attractions.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['itinerary_id'], ['itineraries.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('itinerary_hotels_restaurants',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('itinerary_id', sa.Integer(), nullable=False),
-    sa.Column('hotel_restaurant_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['hotel_restaurant_id'], ['hotels_and_restaurants.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['itinerary_id'], ['itineraries.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('itinerary_transportations',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('itinerary_id', sa.Integer(), nullable=False),
-    sa.Column('transportation_id', sa.Integer(), nullable=False),
-    sa.ForeignKeyConstraint(['itinerary_id'], ['itineraries.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['transportation_id'], ['transportation.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index('idx_activity_district_id', 'activities', ['district_id'], unique=False)
-    op.create_index('idx_attraction_district_id', 'attractions', ['district_id'], unique=False)
-    op.create_index('idx_district_province_id', 'districts', ['province_id'], unique=False)
-    op.create_index('idx_hotels_restaurants_district_id', 'hotels_and_restaurants', ['district_id'], unique=False)
-    op.create_index('idx_transportation_district_id', 'transportation', ['district_id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index('idx_transportation_district_id', table_name='transportation')
-    op.drop_index('idx_hotels_restaurants_district_id', table_name='hotels_and_restaurants')
-    op.drop_index('idx_district_province_id', table_name='districts')
-    op.drop_index('idx_attraction_district_id', table_name='attractions')
-    op.drop_index('idx_activity_district_id', table_name='activities')
-    op.drop_table('itinerary_transportations')
-    op.drop_table('itinerary_hotels_restaurants')
-    op.drop_table('itinerary_attractions')
-    op.drop_table('itinerary_activities')
-    op.drop_table('itineraries')
-    op.drop_table('user_itineraries')
-    # ### end Alembic commands ###
