--- alembic/versions/c00423de709a_added_all_the_category_tables.py
@@ -1,98 +1,0 @@
-"""added all the category tables
-
-Revision ID: c00423de709a
-Revises: 
-Create Date: 2024-09-01 00:09:15.039921
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c00423de709a'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('provinces',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('title')
-    )
-    op.create_table('districts',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('province_id', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['province_id'], ['provinces.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('title')
-    )
-    op.create_table('activities',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('district_id', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['district_id'], ['districts.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('title')
-    )
-    op.create_table('attractions',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('district_id', sa.Integer(), nullable=False),
-    sa.Column('type', sa.String(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['district_id'], ['districts.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('hotels_and_restaurants',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('district_id', sa.Integer(), nullable=False),
-    sa.Column('type', sa.String(), nullable=False),
-    sa.Column('cuisine', sa.String(), nullable=True),
-    sa.Column('comfort', sa.String(), nullable=True),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['district_id'], ['districts.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('title')
-    )
-    op.create_table('transportation',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('district_id', sa.Integer(), nullable=False),
-    sa.Column('type', sa.String(), nullable=False),
-    sa.Column('origin', sa.String(), nullable=False),
-    sa.Column('destination', sa.String(), nullable=False),
-    sa.Column('description', sa.String(), nullable=True),
-    sa.Column('departure', sa.String(), nullable=True),
-    sa.Column('arrival', sa.String(), nullable=True),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['district_id'], ['districts.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('transportation')
-    op.drop_table('hotels_and_restaurants')
-    op.drop_table('attractions')
-    op.drop_table('activities')
-    op.drop_table('districts')
-    op.drop_table('provinces')
-    # ### end Alembic commands ###
